requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-dotenv==1.0.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
gunicorn==21.2.0
requests==2.31.0

# Channels for WebSockets
channels==4.0.0
daphne==4.0.0

# API Documentation
drf-yasg==1.21.7
